+++ cppcache/src/CacheableObjectArray.cpp
@@ -42,7 +42,7 @@ void CacheableObjectArray::fromData(DataInput& input) {
   if (len >= 0) {
     input.read();  // ignore CLASS typeid
     input.read();  // ignore string typeid
-    uint16_t classLen = static_cast<uint16_t >(input.readInt16());
+    uint16_t classLen = input.readInt16();
     input.advanceCursor(classLen);
     CacheablePtr obj;
     for (int32_t index = 0; index < len; index++) {
